requirements.txt
python-telegram-bot==20.0
Flask==2.2.2
requests==2.28.1
